+++ src/render/mesh.cpp
@@ -253,7 +253,7 @@ namespace mbm
 
     bool MESH::setSkeletalAuthoringPalette(const SKELETAL_SHADER_METHOD method,
                                            const float *rows, const uint32_t rowCount,
-                                           const uint32_t *orderedBoneIds, const uint32_t boneIdCount,
+                                           const uint64_t *orderedBoneIds, const uint32_t boneIdCount,
                                            const float time, char *errorOut,
                                            const int errorOutLen) noexcept
     {
